Only forward logs since last boot
This is an attempt to limit the number of log entries that are sent when
the service starts. Without this param it will send all logs in the
history which seems like a bad default. I wish I could make it send only
from when the service starts, but that doesn't seem trivial. I tried
using the `Range:` header as is documented here:
http://www.freedesktop.org/software/systemd/man/systemd-journal-gatewayd.service.html
but it seems to require an actual log entry cursor which doesn't seem
trivial to query for. I could be missing something, but documentation is
sparse for this feature.